Add a modulo for the Add functionality
